refactor: Suspense로 서버, 클라이언트 컴포넌트 분리
